[XENFIN-1237] Move ChipIconButton to its own file
